+++ eventmesh-storage/eventmesh-storage-pulsar/src/main/java/org/apache/eventmesh/storage/pulsar/utils/CloudEventUtils.java
@@ -15,25 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.connector.pulsar.config;
+package org.apache.eventmesh.storage.pulsar.utils;
 
-import org.apache.eventmesh.common.config.Config;
-import org.apache.eventmesh.common.config.ConfigFiled;
+import org.apache.eventmesh.api.SendResult;
 
-import lombok.Getter;
-import lombok.Setter;
+import io.cloudevents.CloudEvent;
 
-@Getter
-@Setter
-@Config(prefix = "eventMesh.server.pulsar", path = "classPath://pulsar-client.properties")
-public class ClientConfiguration {
+public class CloudEventUtils {
 
-    @ConfigFiled(field = "service")
-    private String serviceAddr;
-
-    @ConfigFiled(field = "authPlugin")
-    private String authPlugin;
-
-    @ConfigFiled(field = "authParams")
-    private String authParams;
+    public static SendResult convertSendResult(CloudEvent cloudEvent) {
+        SendResult sendResult = new SendResult();
+        sendResult.setTopic(cloudEvent.getSubject());
+        sendResult.setMessageId(cloudEvent.getId());
+        return sendResult;
+    }
 }
