+++ eventmesh-storage/eventmesh-storage-pulsar/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.consumer.Consumer
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-pulsar=org.apache.eventmesh.connector.pulsar.consumer.PulsarConsumerImpl
+pulsar=org.apache.eventmesh.storage.pulsar.consumer.PulsarConsumerImpl
