+++ eventmesh-storage/eventmesh-storage-pulsar/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.producer.Producer
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-pulsar=org.apache.eventmesh.connector.pulsar.producer.PulsarProducerImpl
+pulsar=org.apache.eventmesh.storage.pulsar.producer.PulsarProducerImpl
